+++ types/components/dv-field.locale.d.ts
@@ -1,0 +1,13 @@
+declare namespace _default {
+    namespace en {
+        let label: string;
+        let error: string;
+    }
+    namespace tr {
+        let label_1: string;
+        export { label_1 as label };
+        let error_1: string;
+        export { error_1 as error };
+    }
+}
+export default _default;