+++ types/components/dv-state.locale.d.ts
@@ -1,0 +1,19 @@
+declare namespace _default {
+    namespace en {
+        let loading: string;
+        let error: string;
+        let retryLabel: string;
+        let empty: string;
+    }
+    namespace tr {
+        let loading_1: string;
+        export { loading_1 as loading };
+        let error_1: string;
+        export { error_1 as error };
+        let retryLabel_1: string;
+        export { retryLabel_1 as retryLabel };
+        let empty_1: string;
+        export { empty_1 as empty };
+    }
+}
+export default _default;